opt: Create separate shareable rule_gen.go file
This prepares for the explorer generator, which will share a new
ruleGen class. This is a separate commit to make it easier to see
changes.

Release note: None
